bq-streaming-webhook/requirements.txt
google-cloud-bigquery==1.20.0
